process documents on the client side and use miniwhoosh endpoints
The processing of a document is now done on the client side using
pdftotext and miniwhoosh. We extract metadata of the documents using
exiftool. The documents contents are then encrypted and sent to the
server where the miniwhoosh index injests the fields we created in the
frontend.
Currently there is no support for downloading the files, oppening a
document or the statistics endpoint.
Also we removed the localization from the app.